#200 [update - user]  회원 탈퇴 시, 비밀번호를 requestbody로 받도록 수정

+++ src/main/java/com/sparta/ticketauction/domain/user/service/impl/UserServiceImpl.java
@@ -11,6 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.sparta.ticketauction.domain.user.entity.User;
 import com.sparta.ticketauction.domain.user.repository.UserRepository;
 import com.sparta.ticketauction.domain.user.request.UserCreateRequest;
+import com.sparta.ticketauction.domain.user.request.UserDeleteRequest;
 import com.sparta.ticketauction.domain.user.request.UserPasswordUpdateRequest;
 import com.sparta.ticketauction.domain.user.request.UserUpdateRequest;
 import com.sparta.ticketauction.domain.user.response.UserResponse;
@@ -113,9 +114,11 @@ public class UserServiceImpl implements UserService {
 
 	@Override
 	@Transactional
-	public void deleteUser(User loginUser, Long userId) {
+	public void deleteUser(User loginUser, Long userId, UserDeleteRequest request) {
 		User user = checkAndGetUser(loginUser, userId);
-
+		if (!passwordEncoder.matches(request.getPassword(), user.getPassword())) {
+			throw new ApiException(NOT_MATCH_PASSWORD);
+		}
 		user.delete();
 	}
